PRO-2997: move subfield_tree.rb and resolved_subfields.rb out of the reflection directory
File moves only, so git records renames and the next commit's diff is the nesting.
