+++ spec/axn/standalone_require_spec.rb
@@ -84,7 +84,7 @@ RSpec.describe "standalone require completeness" do
       ["axn/internal/shape_graph.rb", "Axn::Internal::Reflection::PropertyNames"],
       ["axn/internal/reflection/schema.rb", "Axn::Core::Contract::GENERATED_READER_SOURCE_PATH"],
       ["axn/internal/reflection/schema.rb", "Axn::Validation::Base"],
-      ["axn/internal/reflection/subfield_tree.rb", "Schema"],
+      ["axn/internal/subfield_tree.rb", "Schema"],
     ]
   end
